r5x: configs: Provide Radio HAL on vendor In [1] Google removed many HIDL HAL libraries from VNDK, causing them to only be installed to the system partition.  This is inaccessible for vendor apps and results in the usual:
    linker  : CANNOT LINK EXECUTABLE "/vendor/bin/hw/vendor.somc.hardware.modemswitcher@1.0-service": library "android.hardware.radio.config@1.0.so" not found: needed by /odm/lib64/libril-qc-hal-qmi.so in namespace (default)
    linker  : CANNOT LINK EXECUTABLE "/vendor/bin/hw/qcrild": library "android.hardware.radio.deprecated@1.0.so" not found: needed by /odm/lib64/libril-qc-hal-qmi.so in namespace (default)

These libraries are listed in common-treble because they're used by ODM apps but have no explicit vendor dependency (ie. a binary placed on /vendor) to put them in a vendor-available space.  The default is to include system variants which is why the dependencies have to be suffixed with `.vendor` so that they're shipped in `/vendor` and are available for our ODM blobs even on GSIs.

[1]: https://cs.android.com/android/_/android/platform/hardware/interfaces/+/d610435ac4bb051761b0f016aa6cdf2e884c55b5

Signed-off-by: Marijn Suijten <marijn.suijten@somainline.org>
Change-Id: I8344b85fb4e86a913bd8d9071ce49150ed093575